fix: исправляет вложенные компоненты
